[Fix] 中間データ生成部分のファイルを分割
