+++ chapter01/statement.ts
@@ -1,68 +1,8 @@
+import { createStatementData } from './createStatementData';
+
 export function statement(invoice, plays) {
   const statementData: any = createStatementData(invoice, plays);
   return renderPlainText(statementData, plays);
-
-  function createStatementData(invoice, plays) {
-    const result: any = {};
-    result.customer = invoice.customer;
-    result.performances = invoice.performances.map(enrichPerformance);
-    result.totalAmount = totalAmount(result);
-    result.totalVolumeCredits = totalVolumeCredits(result);
-    return result;
-  }
-
-  function enrichPerformance(aPerformance) {
-    const result: any = { ...aPerformance };
-    result.play = playFor(aPerformance);
-    result.amount = amountFor(result);
-    result.volumeCredits = volumeCreditsFor(result);
-    return result;
-  }
-
-  function playFor(aPerformance) {
-    return plays[aPerformance.playID];
-  }
-
-  function amountFor(aPerformance) {
-    let result = 0;
-    switch (aPerformance.play.type) {
-      case 'tragedy':
-        result = 40000;
-        if (aPerformance.audience > 30) {
-          result += 1000 * (aPerformance.audience - 30);
-        }
-        break;
-      case 'comedy':
-        result = 30000;
-        if (aPerformance.audience > 20) {
-          result += 10000 + 500 * (aPerformance.audience - 20);
-        }
-        result += 300 * aPerformance.audience;
-        break;
-      default:
-        throw new Error(`unknown type: ${aPerformance.play.type}`);
-    }
-    return result;
-  }
-
-  function volumeCreditsFor(aPerformance) {
-    let result = 0;
-    result += Math.max(aPerformance.audience - 30, 0);
-    // 喜劇のときは10人につき、さらにポイントを加算
-    if ('comedy' === aPerformance.play.type)
-      result += Math.floor(aPerformance.audience / 5);
-    return result;
-  }
-
-  function totalAmount(data) {
-    return data.performances
-      .reduce((total, p) => total + p.amount, 0);
-  }
-
-  function totalVolumeCredits(data) {
-    return data.performances
-      .reduce((total, p) => total + p.volumeCredits, 0);
-  }
 }
 
 function renderPlainText(data: any, plays: any) {
